Safari-611.1.10/LayoutTests/media/media-source/media-source-timeoffset-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
RUN(sourceBuffer.timestampOffset = 10)
RUN(sourceBuffer.appendBuffer(samples))
EVENT(updateend)
EXPECTED (bufferedSamples.length == '6') OK
{PTS({10/1 = 10.000000}), DTS({10/1 = 10.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({11/1 = 11.000000}), DTS({11/1 = 11.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({12/1 = 12.000000}), DTS({12/1 = 12.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({13/1 = 13.000000}), DTS({13/1 = 13.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({14/1 = 14.000000}), DTS({14/1 = 14.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({15/1 = 15.000000}), DTS({15/1 = 15.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
END OF TEST
